feat: Add Entity class for managing entities
The `Entity` class is added to manage entities in the codebase. It provides methods for accessing the entity's ID and properties, as well as checking for equality with another entity. This class will be used for various entities in the project.

This commit is part of the ongoing development and complements the recent addition of the `UID` class for generating unique identifiers.