feat: Add ValueObject class for managing value objects
The `ValueObject` class is added to manage value objects in the codebase. It provides a method for checking equality with another value object based on their properties. This class will be used for various value objects in the project.

This commit complements the recent additions of the `DomainError`, `AggregateRoot`, `Entity`, and `UID` classes and is part of the ongoing development of the project.