feat: Add Result type for representing success or failure
The `Result` type is added to represent the outcome of an operation as either a success or a failure. It consists of two subtypes: `Success` and `Failure`. The `Success` subtype contains a `value` property representing the successful result, while the `Failure` subtype contains a `reason` property representing the reason for failure.