feat: Add Duration value object for managing durations
This commit adds the `Duration` value object to the codebase. The `Duration` class represents a duration in terms of days and provides a getter method for accessing the number of days. This value object will be used for managing durations in the inspector management domain.

This commit complements the recent additions of the `SubscriptionLevel` classe and is part of the ongoing development of the project.